requirements.txt
django
mysqlclient
ipython
python-memcached
django-ipware